Removed use of CSSOM for parsed stylesheet output
Known issue with CSSOM library where duplicate rules are ignored. This makes using the particularly library for output unreliable and unable to guarentee parity.

Simply put, we use RegEx for parsing instead of a CSS parse. The RegEx will not be as accurate, but I believe when weighed out against the CSSOM library output, the benefit outweighs the risk until another library is sought which doesn't focus on standards based parsing.